refactor(app): restore layering — agnostic src/app no longer imports react-app
Fixes the three documented layering violations:

- app-inspector.ts (zero-import window-global instrumentation) moves from
  react-app/shell to app/lib where analytics.ts can legally use it; the 8
  react-app consumers now import from app/lib.
- provider-list-query.ts moves from domains/connections to react-app/infra
  (it is shared query infrastructure used by kernel, shell, components, and
  connections — and infra already hosts query-client.ts). Kills the
  kernel -> domains edge in global-sync-provider.
- BrowserPanelTab type moves to app/lib/desktop-types.ts (it describes
  desktop IPC bridge payloads); panel-tab-store re-exports it, so desktop.ts
  no longer reaches into the session domain.

Invariant restored: grep 'react-app/' in src/app -> zero hits.

Tested:
- pnpm typecheck (clean)
- bun test tests/ — 51 pass, 2 pre-existing failures (unchanged vs base)
- pnpm build (vite production build clean)
- live: vite dev server 200, analytics module serves with new import graph
