+++ apps/app/src/app/lib/app-inspector.ts
