+++ apps/app/src/react-app/infra/provider-list-query.ts
@@ -1,8 +1,8 @@
 import { useQuery, type QueryClient } from "@tanstack/react-query";
 
-import type { Client, ModelRef, ProviderListItem } from "../../../app/types";
-import { unwrap } from "../../../app/lib/opencode";
-import { dispatchNewProviders } from "../../../app/lib/provider-events";
+import type { Client, ModelRef, ProviderListItem } from "../../app/types";
+import { unwrap } from "../../app/lib/opencode";
+import { dispatchNewProviders } from "../../app/lib/provider-events";
 import type { ProviderListResponse } from "@opencode-ai/sdk/v2/client";
 
 export const PROVIDER_LIST_CACHE_MS = 5 * 60 * 1000;
